refactor: Update deploy flag in config.yaml files
Update the deploy flag in the config.yaml files for bp-probe-lb-agent and bp-probe-lb to set it to true instead of false. This change is made to enable the deployment of these applications.

Add cluster-wide-cdf.yaml

Add a new file, cluster-wide-cdf.yaml, which contains RBAC configuration for the cluster. This file sets up a cluster role binding for the system:serviceaccount:production:default user to have cluster-admin privileges.