refactor: Update config.yaml to include ignoreDifferences for probe-lb application
This commit updates the config.yaml file for the probe-lb application in the production namespace. It adds the ignoreDifferences configuration for two resources: NetworkAttachmentDefinition from the 'k8s.cni.cncf.io' group and vProbe from the 'topology.radcom.com' group. The ignoreDifferences configuration specifies the jsonPointers for the /spec/config and /spec/customAttributes fields respectively.